[refactor] #149 - 도메인 묶어서 repository, interface 정의
