--- POTI-iOS/POTI-iOS/Domain/Interface/FeedsInterface.swift
@@ -1,10 +1,0 @@
-//
-//  FeedsInterface.swift
-//  POTI-iOS
-//
-//  Created by mandoo on 1/15/26.
-//
-
-protocol FeedsInterface {
-    func fetchFeedsData(artistId: Int?, sort: FeedsSortOption, page: Int) async throws -> FeedsEntity
-}
