--- POTI-iOS/POTI-iOS/Domain/Interface/HomeInterface.swift
@@ -1,10 +1,0 @@
-//
-//  HomeInterface.swift
-//  POTI-iOS
-//
-//  Created by mandoo on 1/15/26.
-//
-
-protocol HomeInterface {
-    func fetchHomeData() async throws -> HomeEntity
-}
