--- POTI-iOS/POTI-iOS/Domain/Interface/PotDetailInterface.swift
@@ -1,10 +1,0 @@
-//
-//  PotDetailInterface.swift
-//  POTI-iOS
-//
-//  Created by mandoo on 1/19/26.
-//
-
-protocol PotDetailInterface {
-    func fetchPotDetail(postId: Int) async throws -> PotDetailEntity
-}
